+++ integrationtests/foldersync/util.go
@@ -1,4 +1,4 @@
-package main
+package foldersync
 
 import (
 	"context"
